Create DatabaseVM
